+++ src/qwertypy/data_plots/plot_colors.py
@@ -1,0 +1,5 @@
+BLACK = "#000000"
+GREEN = "#77b978"
+LIGHT_GREY = "#f7f7f7"
+RED = "#f37262"
+YELLOW = "#fce15d"