+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased (0.5.0)]
+
+### Added
+
+-   `qwertypy.data_plots.trend_plot.trendPlot`.
+
 ## [0.4.0]
 
 ### Added
@@ -22,7 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
--   `qwertypy.tickertape.financials`: Fixed `getStatement`.
+-   `qwertypy.tickertape.financials.getStatement`.
 
 ### Internal